Rename the folder UI to ChangePasswordUI
Rename the folder UI to ChangePasswordUI
